[dashboard, protocol] Extract UserEnvVar.validate() into protocol
